Module move_transactional_test_runner::tasks

Structs

EmptyCommand

InitCommand

PrintBytecodeCommand

Translates the given Move IR module or script into bytecode, then prints a textual representation of that bytecode.

PublishCommand

RunCommand

TaskInput

ViewCommand

Enums

PrintBytecodeInputChoice

When printing bytecode, the input program must either be a script or a module.

SyntaxChoice

TaskCommand

Functions

taskify